Update README.md
Written in Python and R - allows user to impute missing data using a variety of algorithms, and then estimate differential abundance on the corrected matrix. The proteomics matrix is computed in Python using the model of choice among missForest, KNN or RegImpute from the missingpy and StaticImpute libraries. All PCA and Volcano plots are generated in R using Shiny.